tests: share bundle helpers and deduplicate suite scaffolding
Moves bundle discovery into tests/bundles.py, folds the repeated
chunked-prefill check into the llm suite reusing its existing helper,
drops import-time HF env mutation for local_files_only, and extracts
the shared tool-capture setup in the server tests.
